Adds separate actions for endpoints and broken links:
