+++ app/tests/blog/drafts.js
@@ -2,11 +2,11 @@ describe("drafts work", function () {
   const Template = require("models/template");
   const Blog = require("models/blog");
   const sync = require("sync");
-  const blogServer = require("blog");
+  const blogServer = require("../../blog");
   const fs = require("fs-extra");
   const request = require("request");
   const Express = require("express");
-  const config = require("../../config");
+  const config = require("config");
   const guid = require("helper/guid");
 
   global.test.blog();
