+++ tests/unused-dependencies/dependencies.js
